Changed: ``phenomic/lib/PageContainer`` has been relocated.
To prevent issue in the future, is now accessible by doing
``import { PageContainer } from "phenomic"``.
If you want to import it with a different name, you can do it this way:
```js
import { PageContainer as PhenomicPageContainer } from "phenomic"
```
