+++ CHANGELOG.md
@@ -6,6 +6,13 @@
 - Added: 🚀 Hard source Webpack plugin to improve performance by more than 300%
   You can enable this **experimental** feature by add
   ``"webpackHardCache": true`` in your ``package.json`` ``phenomic`` section.
+- Changed: ``phenomic/lib/PageContainer`` has been relocated.
+  To prevent issue in the future, is now accessible by doing
+  ``import { PageContainer } from "phenomic"``.
+  If you want to import it with a different name, you can do it this way:
+  ```js
+  import { PageContainer as PhenomicPageContainer } from "phenomic"
+  ```
 
 ## Boilerplate
